requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
common==0.1.2
common-decorators==0.0.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.1
django-betterforms==2.0.0
django-bootstrap4==22.2
django-heroku==0.3.1
django-taggit==3.0.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.9.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
